agents/exchange_agent/requirements.txt
google-adk==0.1.0
azure-identity
msgraph-sdk
microsoft-kiota-abstractions
requests
python-dotenv
pytz
typing-extensions